\name{pivotPoints-package}
\alias{pivotPoints-package}
\alias{pivotPoints}
\docType{package}
\title{
\packageTitle{pivotPoints}
}
\description{The application of Camarilla pivots in trading}
\packageDescription{pivotPoints}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{pivotPoints}
\packageIndices{pivotPoints}
Use join_data() to put separate .csv files together
USe calcProfit() to test earning of given parameters on certain commodity
}
\author{
\packageAuthor{pivotPoints}

Maintainer: \packageMaintainer{pivotPoints}
}
\references{
\link{http://www.futuresmag.com/2012/10/24/trading-stocks-camarilla-pivots}
}

\keyword{ pivot }
\seealso{
}
\examples{
}
